feat(additional-properties): adds typed expando object schema (#193)
This commit adds a new schema for validating, mapping and un-mapping additional properties in an object. It will also validate and skip the additional properties that do not match the specified type while mapping. While the un-mapping step will fail if the additional properties key matches one of the objects original keys.